OO-374: avoiding multiple udpates
